requirements.txt
chainlit>=0.7.700
openai>=1.1.1
python-dotenv>=1.0.0 